Spring Boot project Rest CRUD API with Hibernate

+++ src/main/java/com/zamoiski/mastery_java/service/EmployeeServiceImpl.java
@@ -2,6 +2,7 @@ package com.zamoiski.mastery_java.service;
 
 import com.zamoiski.mastery_java.dao.EmployeeDAO;
 import com.zamoiski.mastery_java.entity.Employee;
+import com.zamoiski.mastery_java.error.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,8 +12,12 @@ import java.util.List;
 @Service
 public class EmployeeServiceImpl implements EmployeeService {
 
-    @Autowired
     private EmployeeDAO employeeDAO;
+
+    @Autowired
+    public EmployeeServiceImpl(EmployeeDAO employeeDAO){
+        this.employeeDAO=employeeDAO;
+    }
 
     @Override
     @Transactional
@@ -22,8 +27,14 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     @Transactional
-    public Employee findById(int theId) {
-        return employeeDAO.findById(theId);
+    public Employee findById(long theId) {
+        Employee employee=employeeDAO.findById(theId);
+
+        if(employee==null){
+            throw new NotFoundException("Employee is not found - "+ theId);
+        }
+
+        return employee;
     }
 
     @Override
@@ -34,7 +45,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     @Transactional
-    public void deleteById(int theId) {
+    public void deleteById(long theId) {
+        if (employeeDAO.findById(theId)==null){
+            throw new NotFoundException("Employee is not found - "+ theId);
+        }
         employeeDAO.deleteById(theId);
     }
 }
